[test_hw] fix PWM e lettura Encoder

+++ Core/Components/MOTOR/MotorInt.h
@@ -10,10 +10,10 @@
 
 #include "MotorExt.h"
 
-#define PWM_FREQUENCY_KHZ 25	/* f_in = 20 --> 30 KHz */
+#define PWM_FREQUENCY_KHZ 		   25			/* f_in = [20, 30] KHz */
 
-#define ARR_VALUE 1680	/* Value of the ARR Register for source clock frequency 84MHz*/
-#define PSC_VALUE 1	/* Value of the PSC Register */
+#define ARR_VALUE 	     (uint16_t)1680			/* Value of the ARR Register for source clock frequency 84MHz*/
+#define PSC_VALUE 				   1			/* Value of the PSC Register */
